+++ src/main/java/com/elmakers/mine/bukkit/action/builtin/DamageAction.java
@@ -44,7 +44,7 @@ public class DamageAction extends BaseSpellAction
 	public SpellResult perform(CastContext context)
 	{
         Entity entity = context.getTargetEntity();
-		if (entity != null && !(entity instanceof LivingEntity))
+		if (entity == null || !(entity instanceof LivingEntity))
 		{
 			return SpellResult.NO_TARGET;
 		}