+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Added several new tweakshop actions to configure quick-cast and drop/right-click modes
  - Add Cursed Meat item (use it if you dare dare dare dare!)
+ - Add Debugger spell, for debugging spell issues
 
 ## 5.2